+++ src/main/java/com/codenear/butterfly/kakaoPay/application/CancelPaymentService.java
@@ -4,19 +4,15 @@ import com.codenear.butterfly.kakaoPay.domain.CancelPayment;
 import com.codenear.butterfly.kakaoPay.domain.CanceledAmount;
 import com.codenear.butterfly.kakaoPay.domain.OrderDetails;
 import com.codenear.butterfly.kakaoPay.domain.dto.OrderStatus;
-import com.codenear.butterfly.kakaoPay.domain.dto.request.CancelRequestDTO;
 import com.codenear.butterfly.kakaoPay.domain.dto.kakao.CancelResponseDTO;
+import com.codenear.butterfly.kakaoPay.domain.dto.request.CancelRequestDTO;
 import com.codenear.butterfly.kakaoPay.domain.repository.CancelPaymentRepository;
 import com.codenear.butterfly.kakaoPay.domain.repository.OrderDetailsRepository;
+import com.codenear.butterfly.kakaoPay.util.KakaoPaymentUtil;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpHeaders;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.client.RestTemplate;
 
-import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
 
@@ -24,36 +20,17 @@ import java.util.Objects;
 @Transactional
 @RequiredArgsConstructor
 public class CancelPaymentService {
-
-    @Value("${kakao.payment.cid}")
-    private String CID;
-
-    @Value("${kakao.payment.secret-key-dev}")
-    private String secretKey;
-
-    @Value("${kakao.payment.host}")
-    private String host;
-
     private final CancelPaymentRepository cancelPaymentRepository;
     private final OrderDetailsRepository orderDetailsRepository;
+    private final KakaoPaymentUtil<Object> kakaoPaymentUtil;
 
     public void cancelKakaoPay(CancelRequestDTO cancelRequestDTO) {
 
         OrderDetails orderDetails = orderDetailsRepository.findByOrderCode(cancelRequestDTO.getOrderCode());
 
-        Map<String, Object> parameters = new HashMap<>();
-        parameters.put("cid", CID);
-        parameters.put("tid", orderDetails.getTid());
-        parameters.put("cancel_amount", cancelRequestDTO.getCancelAmount());
-        parameters.put("cancel_tax_free_amount", 0);
+        Map<String, Object> parameters = kakaoPaymentUtil.getKakaoPayCancelParameters(orderDetails,cancelRequestDTO);
 
-        HttpEntity<Map<String, Object>> requestEntity = new HttpEntity<>(parameters, this.getHeaders());
-        RestTemplate restTemplate = new RestTemplate();
-
-        CancelResponseDTO cancelResponseDTO = restTemplate.postForObject(
-                host+"/cancel",
-                requestEntity,
-                CancelResponseDTO.class);
+        CancelResponseDTO cancelResponseDTO = kakaoPaymentUtil.sendRequest("/cancel",parameters,CancelResponseDTO.class);
 
         CancelPayment cancelPayment = createCancelPayment(cancelResponseDTO);
 
@@ -90,13 +67,5 @@ public class CancelPaymentService {
         canceledAmount.setPoint(cancelResponseDTO.getAmount().getPoint());
         canceledAmount.setDiscount(cancelResponseDTO.getAmount().getDiscount());
         return canceledAmount;
-    }
-
-    // 카카오가 요구한 헤더값
-    private HttpHeaders getHeaders() {
-        HttpHeaders headers = new HttpHeaders();
-        headers.set("Content-Type", "application/json");
-        headers.set("Authorization", "SECRET_KEY " + secretKey);
-        return headers;
     }
 }
